Levenshtein distance solution (python implementation)
- basic naive (recursive) implementation
- dynamic programming solution (with caching)

Pay attention to the line 7 and 8 we use opposite nested loops in comparison to java implementation since y is bigger then x we might face this array out of bound exception. Probably this is some kind of peculiarities of python you have to work it out why it is so.

Jira:
https://gym1q.atlassian.net/browse/CDX-19